fix: adapting usage of response data
As Barna001 wrote in his review axios is already performing the JSON parsing, so another call of JSON.parse will cause an error. I adapted the code according to the pull request #172. Besides that I named the variablein the same was as #172 as translation is better understandable than just data.

Relates SUITEDEV-29314